pkg/adaptation: add API for runtime integration.
Add a package for interfacing NRI from runtimes, along
with a few additional supporting packages.

This implementation abandons the original exec-based one-
shot plugin design in favour of plugins with daemon-like
lifecycle. These plugins use the new ttRPC-based protocol
to communicate with NRI proper over a unix-domain socket.
A single OS process corresponds to a logical instance of
a plugin. It is responsible for handling the full stream
of requests/events intended for that plugin.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>