pkg/stub: add plugin stub.
Add a plugin stub to handle many of the common, low-level
details of writing an NRI plugin, including
  - connection discovery/establishment
  - connection and state monitoring
  - registering with NRI
  - subscribing for pod/container events

The stub is expected to be the primary API for writing NRI
plugins. The API tries to hide the startup differences of
NRI-launched vs. externally launched plugins. It also aims
to hide the details of ttRPC-based communication so that
other means of plugging in plugins to NRI proper could be
added/experimented with in the future.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>