api: put default socket in nri-specific directory.
Change default socket path to /var/run/nri/nri.sock. This
should work better for plugins which are run as Kubernetes
daemonsets. With the socket directory bind-mounted instead
of the socket itself, plugins should be able to reconnect
if the runtime is ever restarted.

Additionally, create any missing NRI/socket directory with
0700 rights (rwx for owner only) instead of 0755.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>