Implement Pre- and PostSetupNetwork messages
PreSetupNetwork messages are sent to plugins before CNI is invoked
to set up pod networking. The plugins can return a set of additional
CNI capabilities and arguments.
Once CNI networking has completed, PostSetupNetwork messages containing
an equivalent of the CNI Result structure are sent out to interested
plugins, which can return the same or a modified version of the received
CNI Result structure.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>