Implement PreNetworkDeleted and PostNetworkDeleted messages
When the pod sandbox network is to be removed, NRI plugins are informed
of the event in form of PreNetworkDeleted message and a PostNetworkDeleted
event.
The PreNetworkDeleted request is sent to the plugins while the network
namespace is still available and waits until all plugins have responded.
Once all plugins have responded to the PreNetworkDeleted message, the
namespace is removed, and PostNetworkDeleted events will be sent to
plugins subscribed to the event.

Signed-off-by: Patrik Flykt <patrik.flykt@intel.com>