socketpair_windows: remove implementation for now
Current implementation has the same issues that
socketpair_unix.go had, namely double closes and
leaking FDs. The initial author prefers to remove
the code for now as the whole feature (NRI) doesn't
work on windows for now.

This allows us to rename socketpair_unix.go -> socketpair.go
as everything in it is generic.

Signed-off-by: Etienne Champetier <e.champetier@ateme.com>