adaptation: implement plugin synchronization blocks.
Add a mechanism to block and unblock plugin synchronization.
We use this first in our tests to ensure that test plugins
get properly synchronized either before or after we create
a pod or a container, but not halfway through it.

We should also do the same in the corresponding paths of the
NRI runtime integration core where otherwise processing NRI
events concurrent to plugin registration could result in an
inconsistent initial state of a new plugin (missing pods or
containers during sync, or omitted event for the plugin).

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>