network device injector plugin
The network device inject plugin allow to inject network interfaces
that are present in the host to the Pods.

The network interface can be renamed and a network configuration can
be passed.

It is important to differentiate between network interface injection and
CNI, as today, CNI is used for container runtimes to provider the
network configuration, it performs also the creation and the
configuration of the interfaces that are injected into the Pod
namespace and provide as a result some properties like the assiged IPs
that are consumed later by the upstream projects like Kubernetes.

Signed-off-by: Antonio Ojea <aojea@google.com>