api,adaptation,stub: exchange versions during registration.
Exchange NRI versions in use during plugin registration. If we
do not get an NRI version from the runtime (too old peer NRI)
try to infer it using the runtime type and version.

The NRI version in use is discovered using the golang runtime
provided build info. For plugins hosted in the main NRI repo
this does not produce any useful result (because NRI is subject
to a replace directive in the plugins' go.mod). Therefore within
the repo we fall back to using a git-describe generated version.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>