adaptation, plugins: implement default validation.
Implement default (container creation/adjustment) validation
as a builtin plugin. The default validator can be configured
to reject OCI hook injection. Additionally, containers can be
annotated with a set of required plugins. If annotated, these
plugins must be present during container creation or else the
creation of the container is rejected by the validator.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>