ci: add github workflow for publishing container images
Publish multiarch container images of the sample plugins at
ghcr.io/containerd/nri/plugins/<plugin>:<tag>

The tag is the git tag name for tagged versions and "unstable" for the
main branch.

The default-validator and wasm plugins are skipped in the config. The
default-validator plugin is not runnable as an external plugin
and the wasm plugin cannot be run from a container.

Signed-off-by: Markus Lehtonen <markus.lehtonen@intel.com>