Makefile: build proto files unconditionally
Ditch the makefile pattern rule which simply doesn't work in scenarios
where both the source and the build targets are stored in the git repo
(as git operations mangle the time stamps).

So much grey hair avoided when the stuff just force generates
everything instead of leaving you wonder why the tools didn't work as
expected.

Signed-off-by: Markus Lehtonen <markus.lehtonen@intel.com>