api: add v1beta1 wire protocol revision.
Clone v1alpha1 as a starting point of v1beta1. Add dedicated RPC calls
for requests we used to handle all in v1alpha1 with StateChange tagged
with Event's to tell them apart. Remove now unused StateChange.

Likewise, replace usage of Empty as a response type with dedicated per
RPC call reponses. Remove now unused Empty.

Update Shutdown to allow passing a string which describes the reason
for the shutdown of the plugin.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>