api: update v1beta1 protocol definition.
Add dedicated regular RPC calls and messages for requests we used to
all handle in v1alpha1 with StateChange tagged with Event's to tell
them apart. Remove unused StateChange and StateChangeEvent. Replace
Empty similarly as a response type with dedicated reponses specific
to each request. Remove Empty as an unused type.

Update Shutdown, allowing a reason for the shutdown to be passed to
the plugin.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>